Feat. redux 세팅
