+++ .eslintrc.js
@@ -17,6 +17,7 @@ module.exports = {
     "import/no-extraneous-dependencies": "off",
     "react/style-prop-object": "off",
     "react/prop-types": "off",
+    "no-param-reassign": "off",
     "no-use-before-define": "off",
     "react/jsx-props-no-spreading": "off",
     "array-callback-return": "off",
