+++ package.json
@@ -10,6 +10,7 @@
     "lint:fix": "eslint --fix ."
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.3",
     "expo": "~48.0.10",
     "expo-image": "~1.0.0",
     "expo-sensors": "~12.1.1",
@@ -18,7 +19,8 @@
     "matter-js": "^0.19.0",
     "react": "18.2.0",
     "react-native": "0.71.6",
-    "react-native-game-engine": "^1.2.0"
+    "react-native-game-engine": "^1.2.0",
+    "react-redux": "^8.0.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
